OpenCL: Exponents > 2
Now, exponents greater than two may be used for building the result
tree. With an exponent of four (always grouping 16 values), the
implementation is now (for me, at least) about as fast as the CPU
implementation for the original sequence and even faster when exceeding
about 100 MB.

To achieve this, the host pointer policy has to be changed to
CL_MEM_USE_HOST_PTR (leave buffer in host memory).

Measurements are taken for the index being half of the sequence length
(which is around two thirds into the sequence).

There is perhaps still room for improvement regarding the generated
result buffer (it is still too big).